pyproject.toml
[project]
name = "fastapi-oauth-rbac"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Fernando", email = "fernando@softpoint.es" }
]
requires-python = ">=3.14"
dependencies = [
    "aiosqlite>=0.22.1",
    "alembic>=1.18.3",
    "argon2-cffi>=25.1.0",
    "email-validator>=2.2.0",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "pwdlib[argon2]>=0.3.0",
    "pydantic-settings>=2.12.0",
    "pytest-asyncio>=1.3.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "sqlalchemy[asyncio]>=2.0.46",
    "uvicorn>=0.34.0",
]

[build-system]
requires = ["uv_build>=0.9.15,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
]
